% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_Plot.R
\name{bulkCNVs}
\alias{bulkCNVs}
\title{Arm-level copy number variations of bulk DNA-seq data.}
\usage{
bulkCNVs(centmere, dcnr, dcns)
}
\arguments{
\item{centmere}{the simplified 3-column hg38 centromere data}

\item{dcnr}{cnr file which is cnvkit output}

\item{dcns}{segment file which is cnvkit output}
}
\value{
An arm-level CNV plot, because the ref pool is male, so the plot also
has chrY data for female but the plot shows a deep deletion of chrY for female
}
\description{
Arm-level copy number variations of bulk DNA-seq data.
}
\examples{
\dontrun{
centm <- read.csv(system.file("extdata/", "hg38_centromereSimple.bed", package = "stmut"),
  sep = "\t", header = FALSE
)
cnr <- read.table(system.file("extdata/", "P6_T.deduplicated.realign.cnr",
  package = "stmut"
), sep = "\t", header = TRUE)
cns <- read.table(system.file("extdata/", "P6_T.deduplicated.realign.cns",
  package = "stmut"
), sep = "\t", header = TRUE)
bulkCNVs(centmere = centm, dcnr = cnr, dcns = cns)
}
}
