% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVsScorPermutation.R
\name{caseCNVscore}
\alias{caseCNVscore}
\title{Generate your sample CNV score dataframe.}
\usage{
caseCNVscore(cdt0, armGens, arm, gainLoss)
}
\arguments{
\item{cdt0}{sample cdt file}

\item{armGens}{dataframe of the output of CtArmGenes function}

\item{arm}{A vector of arm with gene gain and loss}

\item{gainLoss}{A vector of data showing gain(+1) and loss(-1)}
}
\value{
A dataframe of your sample CNV score
}
\description{
Generate your sample CNV score dataframe.
}
\examples{
\dontrun{
print("check user guider")
}
}
