% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{cdt}
\alias{cdt}
\title{Sample Data for CtArmGenes and cdt_filt_sort Function}
\format{
A data frame with 1506 rows and 5 variables
\describe{
\item{CLID}{CLID}
\item{NAME}{Gene chr:start-end:name}
\item{spot19_rep1_cluster1}{first spot}
\item{spot424_rep1_cluster1}{second spot}
\item{spot325_rep1_cluster1}{third spot}
}
}
\usage{
cdt
}
\description{
First input for CtArmGenes and cdt_filt_sort
}
\keyword{datasets}
