% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_Plot.R
\name{cdt_filt_sort}
\alias{cdt_filt_sort}
\title{Rank the spots by their similarity to the DNA-seq copy number alterations.}
\usage{
cdt_filt_sort(cdt, genes, gainLoss, rs)
}
\arguments{
\item{cdt}{cdt file to be ranked}

\item{genes}{Number of genes in each arm}

\item{gainLoss}{Gain or Loss info of each arm, 1 represent gain, -1 represent loss.}

\item{rs}{Representative row position of each arm.}
}
\value{
A spots dataframe sorted by CNVs
}
\description{
Rank the spots by their similarity to the DNA-seq copy number alterations.
}
\examples{
\dontrun{
cdt <- read.table(system.file("extdata/", "cdt.cdt", package = "stmut"), header = TRUE)
data4 <- read.csv(system.file("extdata/", "CtArmGenSummary.csv", package = "stmut"), header = TRUE)
arm <- c(
  "1p", "3p", "3q", "4q", "5q", "8q", "9q", "10p", "10q", "11q", "13p", "13q",
  "20p", "20q", "21q", "14q", "17q"
)
d4 <- data4 \%>\% filter(arms \%in\% arm)
genes <- d4[, "genes"]
rs <- d4[, "gene_row"]
gainLoss <- c(1, -1, 1, -1, -1, 1, 1, -1, -1, 1, -1, -1, 1, 1, -1, 1, 1)
df7 <- cdt_filt_sort(cdt = cdt, genes = genes, gainLoss = gainLoss, rs = rs)
}
}
