% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{centm}
\alias{centm}
\title{Sample Data for wtArmMedianOne Function}
\format{
A data frame with 24 rows and 3 variables
\describe{
\item{V1}{Cchromosome column}
\item{V2}{the start position}
\item{V3}{the end position}
}
}
\usage{
centm
}
\description{
Second input for wtArmMedianOne
}
\keyword{datasets}
