% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{cnr}
\alias{cnr}
\title{Sample Data for wtArmMedianOne Function}
\format{
A data frame with 1516 rows and 9 variables
\describe{
\item{chromosome}{the chromosome information}
\item{start}{the start position}
\item{end}{the end position}
\item{gene}{the gene name}
\item{log2}{log2 value}
\item{depth}{depth}
\item{gc}{gc content}
\item{tx_length}{tx_length}
\item{weight}{weight}
}
}
\usage{
cnr
}
\description{
First input for wtArmMedianOne
}
\keyword{datasets}
