% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVsScorPermutation.R
\name{CNVpermut}
\alias{CNVpermut}
\title{Generate CNVs_Scores of 100 permutations.}
\usage{
CNVpermut(m = 100, cdt0, armGens, arm, gainLoss)
}
\arguments{
\item{m}{number of times of permutation, by default is 100}

\item{cdt0}{cdt file used to calculate CNVs}

\item{armGens}{dataframe of the output of CtArmGenes function}

\item{arm}{A vector of arm with gene gain and loss}

\item{gainLoss}{A vector of data showing gain(+1) and loss(-1)}
}
\value{
A dataframe of m times permutation values.
}
\description{
Generate CNVs_Scores of 100 permutations.
}
\examples{
\dontrun{
print("check user guider")
}
}
