% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_Plot.R
\name{cnvPlot1}
\alias{cnvPlot1}
\title{Title DNA bulk data is provided CNVplot1, output plot based on bulk DNA-- CNV data.}
\usage{
cnvPlot1(c7 = c7, data2 = data2, arm = arm, gainLoss = gainLoss)
}
\arguments{
\item{c7}{Internal CNVmatrix}

\item{data2}{The summary.txt file in the wtcnr folder}

\item{arm}{A list of arms with CNV changes, ex: 3p,3q,5p,7q,...}

\item{gainLoss}{A list of numeric value indicate CNV gain or loss, ex: -1,1,1,-1,...}
}
\value{
output 5 files: CNVs_RankedBySimilarityToDNA.cdt,CNVs_RankedBySimilarityToDNA_CNVscoreHistogram.csv,
CNVs_RankedBySimilarityToDNA_CNVscoreHistogram.pdf,CNVs_RankedBySimilarityToDNA_QQplot.pdf,
CNVs_RankedbySimilaritytoDNA_Quintiles4Loupe.csv
}
\description{
Title DNA bulk data is provided CNVplot1, output plot based on bulk DNA-- CNV data.
}
