% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_Plot.R
\name{cnvPlot2}
\alias{cnvPlot2}
\title{Title: Sort CNV first by clusters and then total reads; generating clusters heatmap and chromosome heatmaps.}
\usage{
cnvPlot2(c6 = c6, data3 = data3)
}
\arguments{
\item{c6}{Internal CNV matrix}

\item{data3}{grouped_spotSummary.csv file}
}
\value{
Two files:CNVs_OrganizedByGEcluster_UMIcount.cdt and CNVs_OrganizedByGEcluster_UMIcount.pdf
}
\description{
Title: Sort CNV first by clusters and then total reads; generating clusters heatmap and chromosome heatmaps.
}
