% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_Plot.R
\name{cnvPlot3}
\alias{cnvPlot3}
\title{Title}
\usage{
cnvPlot3(c7, ncluster, method1 = "canberra", method2 = "complete", ...)
}
\arguments{
\item{c7}{Preprocessed cdt0, internal argument}

\item{ncluster}{Number of cluster you want to generate}

\item{method1}{Inherit from heatmap.2x}

\item{method2}{Inherit from heatmap.2x}

\item{...}{Any arguments from heatmap.2x}
}
\value{
Output 2 files: CNVs_clustered.Rdata and CNVs_clustered_heatmap.pdf
}
\description{
Title
}
