% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_Plot.R
\name{cnvs}
\alias{cnvs}
\title{Generate filtered, centered cdt file for spot CNVs or cell CNVs.}
\usage{
cnvs(
  data1,
  cdt,
  data2,
  data3,
  ncluster = 4,
  arm = c(),
  gainLoss = c(),
  method1 = "canberra",
  method2 = "complete",
  ...
)
}
\arguments{
\item{data1}{A two-column csv file; the 2 column names are "spotName", "annotate".
In the "annotate" column, spots which are used as reference should be annotated as "normal". These spots are used to center the data.}

\item{cdt}{Original cdt file to be processed.}

\item{data2}{Arm-based gene summary file, output of wtArmMedianOne()}

\item{data3}{The spotSummary file, generated by sptBClstRds()}

\item{ncluster}{By default, it generates 4 clusters heatmap}

\item{arm}{Optional. If you don't have DNAseq data. arm is not required. A vector of arms with CNVs signals}

\item{gainLoss}{Optional. If you don't have DNAseq data. gainLoss is not required. CNVs signals corresponding to arms, -1 is loss, +1 is gain}

\item{method1}{The distance measure to be used. Check ?dist to see more options.}

\item{method2}{The agglomeration method to be used. Check ?hclust to see more options}

\item{...}{Inherit all arguments from heatmap.2x}
}
\value{
Five files: CNVs_clustered_heatmap.pdf,CNVs_OrganizedByGEcluster_UMIcount_ClusterTemplate.png,CNVs_OrganizedByGEcluster_UMIcount_ChromosomeTemplate.png,CNVs_OrganizedByGEcluster_UMIcount.cdt, CNVs_RankedBySimilarityToDNA.cdt.
the CNVs_RankedBySimilarityToDNA.cdt will be generated only when bulk tumor DNA data is provided.
}
\description{
Generate filtered, centered cdt file for spot CNVs or cell CNVs.
}
