% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_Plot.R
\name{cnvsorted}
\alias{cnvsorted}
\title{Generate a cdt file of sorted CNVs data of spots based on DNAseq CNV signals.}
\usage{
cnvsorted(data1, cdt, data2, arm, gainLoss, tumorclusters)
}
\arguments{
\item{data1}{new spots summary data}

\item{cdt}{CNVs data to be sorted.}

\item{data2}{arm-based gene summary file, output of wtArmMedianOne()}

\item{arm}{a vector of arms with CNVs signals}

\item{gainLoss}{CNVs signals corresponding to arms, -1 is loss, +1 is gain}

\item{tumorclusters}{customer-defined tumor clusters.}
}
\value{
a cdt file for JavaTree vies or R visualization.
}
\description{
Generate a cdt file of sorted CNVs data of spots based on DNAseq CNV signals.
}
\examples{
\dontrun{
print("check user guide")
}
}
