% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_Plot.R
\name{CtArmGenes}
\alias{CtArmGenes}
\title{Reformat Gene Summary of Each Arm for cdt_filt_sort function use in the next step.}
\usage{
CtArmGenes(cdt, data)
}
\arguments{
\item{cdt}{the filtered cdt file.}

\item{data}{A summary file generated by wtArmMedianOne function.}
}
\value{
A dataframe prepared for sorting the cdt based on bulk CNVs information.
}
\description{
Reformat Gene Summary of Each Arm for cdt_filt_sort function use in the next step.
}
\examples{
\dontrun{
cdt <- read.table(system.file("extdata/", "cdt.cdt", package = "stmut"), header = TRUE)
data3 <- read.table(system.file("extdata/", "summary.txt", package = "stmut"),
  sep = "\t", header = TRUE
)
df6 <- CtArmGenes(cdt = cdt, data = data3)
}
}
