% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{d1}
\alias{d1}
\title{Sample Data for nonZeRdCts Function}
\format{
A data frame with 10 rows and 6 variables
\describe{
\item{barcode}{the spatial barcode of spots}
\item{cluster}{the cluster the spot belongs to}
\item{spot}{the spot name}
\item{TotalRDs}{the total number of reads of each spot}
\item{array_row}{spot row position}
\item{array_col}{spot col position}
}
}
\usage{
d1
}
\description{
First input for nonZeRdCts
}
\keyword{datasets}
