% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{d2}
\alias{d2}
\title{Sample Data for nonZeRdCts Function}
\format{
A data frame with 5 rows and 6 variables
\describe{
\item{spot}{the spot name}
\item{RreadC}{reference read counts}
\item{MreadC}{Mutant reads counts}
\item{Treads}{the sum of RreadC and MreadC}
\item{groups}{if MreadC == 0, the spot will given "G0", if MreadC == 1, the spot will given "G1",
.... if MreadC > 9, the spot will given "G10"}
\item{barcode}{the spatial barcode of spots}
}
}
\usage{
d2
}
\description{
Second input for nonZeRdCts
}
\keyword{datasets}
