% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{data1}
\alias{data1}
\title{Sample Data for sptBClstRds Function}
\format{
A data frame with 36601 rows and 11 variables
\describe{ ## document each column/variable; no "" is needed for the variable name, otherwise it will cause warning.
\item{X}{the first colum is the gene name}
\item{AAACAAGTATCTCCCA.1}{spatial barcode of spot1}
\item{AGATACTCAAGATCGA.1}{spatial barcode of spot2}
\item{AGATCTCAGGTGTGAT.1}{spatial barcode of spot3}
\item{AGATGACTCGCCCACG.1}{spatial barcode of spot4}
\item{AGCAACATATCTTATT.1}{spatial barcode of spot5}
\item{CAGAACTTAGCCCTCT.1}{spatial barcode of spot6}
\item{CCGTAGGGTTGTTTAC.1}{spatial barcode of spot7}
\item{CCTCCTGTTGTGTCGT.1}{spatial barcode of spot8}
\item{TGATGTCAATTAAGTG.1}{spatial barcode of spot9}
\item{TTACACGATCTGCGAC.1}{spatial barcode of spot10}
}
}
\usage{
data1
}
\description{
Second input for sptBClstRds
}
\keyword{datasets}
