% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{data10}
\alias{data10}
\title{Sample Data for proposTumLoup Function}
\format{
A data frame with 744 rows and 3 variables
\describe{
\item{barcode}{spot barcode}
\item{spot}{spot}
\item{TotalRDs}{TotalRDs}
}
}
\usage{
data10
}
\description{
Second input for proposTumLoup
}
\keyword{datasets}
