% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{data2}
\alias{data2}
\title{Sample Data for sptBClstRds Function}
\format{
A data frame with 10 rows and 2 variables
\describe{
\item{Barcode}{the spatial barcode of spots}
\item{Graph.based}{the cluster of 10X Loupe}
}
}
\usage{
data2
}
\description{
Third input for sptBClstRds
}
\keyword{datasets}
