% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{data3}
\alias{data3}
\title{Sample Data for CtArmGenes Function}
\format{
A data frame with 23 rows and 5 variables
\describe{
\item{chromosome}{the chromosome information}
\item{p_Genes}{number of genes in p arm}
\item{q_Genes}{number of genes in q arm}
\item{pArmEnds}{position pArm ends}
\item{CM_row_pos}{which row the centromere is from}
}
}
\usage{
data3
}
\description{
First input for CtArmGenes
}
\keyword{datasets}
