% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{data4}
\alias{data4}
\title{Sample Data for cdt_filt_sort Function}
\format{
A data frame with 23 rows and 5 variables
\describe{
\item{arms}{arm name}
\item{genes}{number of genes in each arm}
\item{chromosome}{chromosome}
\item{gene_row}{representative gene row}
}
}
\usage{
data4
}
\description{
First input for cdt_filt_sort
}
\keyword{datasets}
