% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{data5}
\alias{data5}
\title{Sample Data for accStartCNR_CNS Function}
\format{
A data frame with 389448 rows and 7 variables
\describe{
\item{chromosome}{chromosome name}
\item{start}{start position of DNA region}
\item{end}{end position of DNA region}
\item{log2}{log2 ratio}
\item{weight}{gene weight}
}
}
\usage{
data5
}
\description{
First input for accStartCNR_CNS
}
\keyword{datasets}
