% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{data6}
\alias{data6}
\title{Sample Data for accStartCNR_CNS Function}
\format{
A data frame with 122 rows and 8 variables
\describe{
\item{chromosome}{chromosome name}
\item{start}{start position of DNA region}
\item{end}{end position of DNA region}
\item{log2}{log2 ratio}
\item{weight}{gene weight}
}
}
\usage{
data6
}
\description{
Second input for accStartCNR_CNS
}
\keyword{datasets}
