% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{data7}
\alias{data7}
\title{Sample Data for accumStartPos Function}
\format{
A data frame with 13401 rows and 4 variables
\describe{
\item{chromosome}{chromosome name}
\item{start}{start position of DNA region}
\item{end}{end position of DNA region}
\item{tumorshift}{tumorshift deviation from 0.5 VAF}
}
}
\usage{
data7
}
\description{
First input for accumStartPos
}
\keyword{datasets}
