% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{data8}
\alias{data8}
\title{Sample Data for groupSpots and newSptBC Function}
\format{
A data frame with 389 rows and 5 variables
\describe{
\item{barcode}{spot barcode}
\item{cX}{X-axis position}
\item{cY}{Y-axis position}
\item{clusters}{the cluster the spot belongs to}
\item{reads}{the number of reads in the spot}
}
}
\usage{
data8
}
\description{
First input for groupSpots
}
\keyword{datasets}
