% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{data9}
\alias{data9}
\title{Sample Data for proposTumLoup Function}
\format{
A data frame with 67 rows and 15 variables
\describe{
\item{barcode}{spot barcode}
\item{TotalRDs}{TotalRDs}
\item{spot}{spot}
\item{RreadC}{RreadC}
\item{MreadC}{MreadC}
\item{Treads}{Treads}
\item{groups}{groups}
\item{propM2Total}{propM2Total}
\item{propM2SumRM}{propM2SumRM}
\item{MoreThan1Mrd}{MoreThan1Mrd}
\item{Score}{Score}
\item{GenesWithMutRead}{GenesWithMutRead}
\item{array_row}{array_row}
\item{array_col}{array_col}
\item{cluster}{cluster}
}
}
\usage{
data9
}
\description{
First input for proposTumLoup
}
\keyword{datasets}
