% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{files}
\alias{files}
\title{Sample Data for sptBClstRds Function}
\format{
A vector containing all spotIndex path
\describe{
\item{files}{a character}
}
}
\usage{
files
}
\description{
First input for sptBClstRds
}
\keyword{datasets}
