% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{files2}
\alias{files2}
\title{Sample Data for sptMutCt Function}
\format{
A vector containing all spots samtools mpileup output path
\describe{
\item{files2}{a character}
}
}
\usage{
files2
}
\description{
Second input for sptBClstRds
}
\keyword{datasets}
