% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_Plot.R
\name{groupSpots}
\alias{groupSpots}
\title{Grouping spots from the same cluster to generate new spots to save spots with few reads.}
\usage{
groupSpots(data, Treads = 5000, NumSpt = 8)
}
\arguments{
\item{data}{A dataframe of spots to be grouped}

\item{Treads}{The min reads of the a new spot, by default Treads=5000}

\item{NumSpt}{NumSpt*NumSpt nearest spots will be included to determine which
spots to grouped together. By default, NumSpt=6}
}
\value{
A dataframe of new spots
}
\description{
Grouping spots from the same cluster to generate new spots to save spots with few reads.
}
\examples{
\dontrun{
clst1 <- read.csv(system.file("extdata/", "c1.csv", package = "stmut"), header = TRUE)
df10 <- groupSpots(clst1, Treads = 5000, NumSpt = 8)
}
}
