% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{index}
\alias{index}
\title{Sample Data for sptMutCt Function}
\format{
A data frame with 10 rows and 3 variables
\describe{
\item{barcode}{the spatial barcode of spots}
\item{spot}{the spot name}
\item{TotalRDs}{the total number reads of each spot}
}
}
\usage{
index
}
\description{
First input for sptMutCt
}
\keyword{datasets}
