% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicImbalance.R
\name{lohMajorAlleleCt}
\alias{lohMajorAlleleCt}
\title{Combine all spot major/minor allele counts into one table.}
\usage{
lohMajorAlleleCt(files, y = 12)
}
\arguments{
\item{files}{A list of pathes of spot mpileupOutput of major- and minor allele.}

\item{y}{The sequential order of the spot name is in the path}
}
\value{
a list of dataframe, the first is all spots SNPs major- and minor- count;
the other is the filtered version of SNPmajorAlleleCount table
}
\description{
Combine all spot major/minor allele counts into one table.
}
\examples{
\dontrun{
print("check user guide")
}
}
