% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicImbalance.R
\name{lohMpileupInput}
\alias{lohMpileupInput}
\title{Create samtools_Mpileup_input for counting spot allelic imbalance.}
\usage{
lohMpileupInput(data1)
}
\arguments{
\item{data1}{Tumor DNAseq germline SNPs ref/Alt counts}
}
\value{
A dataframe serves as the input for samtools mpileup
}
\description{
Create samtools_Mpileup_input for counting spot allelic imbalance.
}
\examples{
\dontrun{
print("check user-guide")
}
}
