% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_Plot.R
\name{newSptBC}
\alias{newSptBC}
\title{Generating new_spots_feature_bc.csv after calling groupSpots function.}
\usage{
newSptBC(df1, Treads = 5000, NumSpt = 8, data)
}
\arguments{
\item{df1}{dataframe of spots to be grouped}

\item{Treads}{min reads of new spot, default is 5000}

\item{NumSpt}{number of spots to be considered, default is 8}

\item{data}{Original filtered_feature_bc.csv}
}
\value{
A list of 2 dataframe and a vector of representative barcodes of new spots.
One is grouping spots with grouping-name; the other is newspots(grouped_spots + non_group_spots) feature barcode dataframe
}
\description{
Generating new_spots_feature_bc.csv after calling groupSpots function.
}
\examples{
\dontrun{
clst1 <- read.csv(system.file("extdata/", "c1.csv", package = "stmut"), header = FALSE)
c1BC <- read.csv(system.file("extdata/", "c1_feature_bc.csv", package = "stmut"), header = FALSE)
df11 <- newSptBC(df1 = clst1, Treads = 5000, NumSpt = 8, data = c1BC)
}
}
