% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointMutations.R
\name{nonZeRdCts}
\alias{nonZeRdCts}
\title{Generate Non-zero Reads Spot Comprehensive Table}
\usage{
nonZeRdCts(df1, df2)
}
\arguments{
\item{df1}{first input: spotBClsterRds.csv}

\item{df2}{second input: spotRdCtFinal.csv}
}
\value{
A list of two dataframes
}
\description{
Generate Non-zero Reads Spot Comprehensive Table
}
\examples{
\dontrun{
d1 <- read.csv(system.file("extdata/", "spotBClster.csv", package = "stmut"), header = TRUE)
d2 <- read.csv(system.file("extdata/", "spotRdCtFinal.csv", package = "stmut"), header = TRUE)
df3 <- nonZeRdCts(df1 = d1, df2 = d2)
}
}
