% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{path}
\alias{path}
\title{Sample Data for sptBClstRds Function}
\format{
A path to tissue_positions_list.csv
\describe{
\item{path}{a character}
}
}
\usage{
path
}
\description{
Fourth input for sptBClstRds
}
\keyword{datasets}
