% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{path1}
\alias{path1}
\title{Sample Data for spotSummary Function}
\format{
A path to spot Mpileup file
\describe{
\item{path}{a character}
}
}
\usage{
path1
}
\description{
Second input for spotSummary
}
\keyword{datasets}
