% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVsScorPermutation.R
\name{permutScore}
\alias{permutScore}
\title{Generate patient scores and permutation scores and rank them from high to low.}
\usage{
permutScore(sptScor, permuts)
}
\arguments{
\item{sptScor}{This is the output of caseCNVscore function.}

\item{permuts}{This is the output of CNVpermut function.}
}
\value{
A list of 2 dataframes. The first is the sample_score_summary,
the second is the permuts score summary.
}
\description{
Generate patient scores and permutation scores and rank them from high to low.
}
\examples{
\dontrun{
print("check user guider")
}
}
