% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_Plot.R
\name{quintilePlot}
\alias{quintilePlot}
\title{Generate quintile plot for Loupe visualization}
\usage{
quintilePlot(df = df, df1 = df1, dat1 = dat1)
}
\arguments{
\item{df}{dataframe from CNVs_RankedBySimilarityToDNA_CNVscoreHistogram.csv}

\item{df1}{dataframe from BarcodeLegend.csv}

\item{dat1}{dataframe from grouped_spotSummary.csv}
}
\value{
A quintile csv file for Loupe visualization
}
\description{
Generate quintile plot for Loupe visualization
}
