% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicImbalance.R
\name{SNPmajorAlleleCt}
\alias{SNPmajorAlleleCt}
\title{Count the major allele and minor allele per mutant of each spot.}
\usage{
SNPmajorAlleleCt(x, y = 12)
}
\arguments{
\item{x}{the path to the mpileup output of major- and minor- allele per mut of each spot}

\item{y}{a numeric value showing the sequential order of the spot name in the path x}
}
\value{
a dataframe of mutant major- and minor- allele counts. one row per mutant per spot
}
\description{
Count the major allele and minor allele per mutant of each spot.
}
\examples{
\dontrun{
print("check user-guide")
}
}
