% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointMutations.R
\name{spaPointMutation}
\alias{spaPointMutation}
\title{Generate combined spatial spot point mutations list.}
\usage{
spaPointMutation(files1, files2, data1, data2, path, path1)
}
\arguments{
\item{files1}{A list of spotIndex files}

\item{files2}{A list of spot mutation list files, samtools mpileup output}

\item{data1}{filtered_feature_bc_.csv, output of spaceranger pipeline}

\item{data2}{Graph-based.csv}

\item{path}{path of tissue_positions_list.csv}

\item{path1}{path of mpileup outputs}
}
\value{
A pointMutation folder includes 8 files,AllSptTumPropsed.csv, spotMutReadCount.csv,
MutantSpotIndex.csv, spotRdCtFinal.csv, NonZeroRdSpotIndex.csv, sptBcRds.csv,
MutantSpotIndex_Scored.csv, sptBcSummary.csv
}
\description{
Generate combined spatial spot point mutations list.
}
\examples{
\dontrun{
print("check user guide")
}
}
