% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointMutations.R
\name{splitSpot}
\alias{splitSpot}
\title{Split filtered_feature_bc.csv file to individual spots.}
\usage{
splitSpot(file)
}
\arguments{
\item{file}{filtered_feature_bc.csv from spaceranger}
}
\value{
Two folders of txt files, spot barcodes and spot gene exp profile
}
\description{
Split filtered_feature_bc.csv file to individual spots.
}
\examples{
\dontrun{
print("check user guide")
}
}
