% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_Plot.R
\name{spotsGrouped}
\alias{spotsGrouped}
\title{Create grouped spots summary files.}
\usage{
spotsGrouped(df1, Treads = 5000, NumSpt = 8, data, locpath)
}
\arguments{
\item{df1}{spots summary file generated by sptBClstRds()}

\item{Treads}{minimum reads for new spot}

\item{NumSpt}{spots range for grouping}

\item{data}{filtered_feature_bc.csv from spaceranger.}

\item{locpath}{path to tissue_positions_list.csv file from spaceranger.}
}
\value{
generate CNVs/spotIndex/, CNVs/txt/ folder; and grpGraph_based.csv,spotGroupingSummary.csv,
newSpotBcRds.csv, newSpotSummary.csv, tissue_positions_list.csv, newSptsGenExp.csv files.
}
\description{
Create grouped spots summary files.
}
\examples{
\dontrun{
print("check user guide")
}
}
