% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointMutations.R
\name{spotSummary}
\alias{spotSummary}
\title{Rank Spots with Mutant Reads using A Customized Strategy}
\usage{
spotSummary(df, path1)
}
\arguments{
\item{df}{A dataframe of nonZeroRdSpotIndex.csv}

\item{path1}{A path to where spot mutant reads statistic file is located}
}
\value{
A comprehensive table mutant information of each spot
}
\description{
Rank Spots with Mutant Reads using A Customized Strategy
}
\examples{
\dontrun{
df <- read.csv(system.file("extdata/", "NonZeroRdSpotIndex.csv", package = "stmut"), header = TRUE)
path1 <- system.file("extdata/mpileup", package = "stmut")
df4 <- spotSummary(df = df, path1 = path1)
}
}
