% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointMutations.R
\name{sptBClstRds}
\alias{sptBClstRds}
\title{Visualize Somatic Alterations of fresh frozen Spatial Transcriptomics Data}
\usage{
sptBClstRds(files, data1, data2, path)
}
\arguments{
\item{files}{A vector containing file path of spatial barcode txt files}

\item{data1}{filtered_feature_bc.csv generated by spaceranger}

\item{data2}{Graph-Based csv file exported from 10X Loupe Browser when open Loupe.Loupe}

\item{path}{Path to the tissue_positions_list.csv under spatial/ folder}
}
\value{
A list of two dataframe, the first dataframe includes spot-barcode-reads,
the second dataframe is used for grouping spots if necessary.
}
\description{
Visualize Somatic Alterations of fresh frozen Spatial Transcriptomics Data
}
\examples{
\dontrun{
files <- list.files(
  path = system.file("extdata/spotIndex", package = "stmut"),
  pattern = ".txt", full.names = TRUE, recursive = FALSE
)
data1 <- read.csv(system.file("extdata/", "filtered_feature_bc.csv",
  package = "stmut"
), header = TRUE)
data2 <- read.csv(system.file("extdata/", "Graph-Based.csv", package = "stmut"), header = TRUE)
path <- system.file("extdata/", "tissue_positions_list.csv", package = "stmut")
df <- sptBClstRds(files = files, data1 = data1, data2 = data2, path = path)
}
}
