% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointMutations.R
\name{sptMutCt}
\alias{sptMutCt}
\title{Count Ref and Mutant Reads for Each Spot}
\usage{
sptMutCt(index, files)
}
\arguments{
\item{index}{A Dataframe of spatial barcode}

\item{files}{Vector of paths of samtools Mpileup output of each spot}
}
\value{
A list of two dataframe
}
\description{
Count Ref and Mutant Reads for Each Spot
}
\examples{
\dontrun{
index <- read.csv(system.file("extdata/", "spotBC.csv", package = "stmut"), header = TRUE)
files2 <- list.files(
  path = system.file("extdata/mpileup", package = "stmut"),
  pattern = "MpileupOutput_RNA.txt", full.names = TRUE, recursive = TRUE, include.dirs = TRUE
)
df1 <- sptMutCt(index = index, files = files2)
}

}
