% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_Plot.R
\name{wtArmMedianOne}
\alias{wtArmMedianOne}
\title{Calculate Chromosome Arm CNVs Weighted Median to Represent the CNVs of that arm.}
\usage{
wtArmMedianOne(data, centmere)
}
\arguments{
\item{data}{The cnr dataframe of each spot.}

\item{centmere}{The centromere dataframe of hg38.}
}
\value{
A list of 2 dataframe, the weighted cnr file and a summary file counting
how many genes in each arm.
}
\description{
Calculate Chromosome Arm CNVs Weighted Median to Represent the CNVs of that arm.
}
\examples{
\dontrun{
cnr <- read.table(system.file("extdata/", "spot1_rep1.cnr", package = "stmut"), header = TRUE)
centm <- read.csv(system.file("extdata/", "hg38_centromereSimple.bed", package = "stmut"),
  sep = "\t", header = FALSE
)
df5 <- wtArmMedianOne(data = cnr, centmere = centm)
}
}
