% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{data8}
\alias{data8}
\title{Sample Data for groupSpots and newSptBC Function}
\format{
A data frame with 389 rows and 6 variables
\describe{
\item{barcode}{spot barcode}
\item{cluster}{the cluster the spot belongs to}
\item{spot}{spot name}
\item{TotalRDs}{the number of reads in the spot}
\item{array_row}{X-axis position}
\item{array_col}{Y-axis position}
}
}
\usage{
data8
}
\description{
First input for groupSpots
}
\keyword{datasets}
